integrated with razorpay

+++ src/main/java/com/ecommerce/KiranaKart/dto/ProductDto.java
@@ -10,7 +10,7 @@ import java.time.LocalDate;
 public class ProductDto {
     private String name; // Name of the product
     private String description; // Description of the product
-    private BigDecimal price; // Price of the product
+    private Double price; // Price of the product
     private int quantity; // Quantity in stock
     private long category_id; // Category of the product
     private String brand; // Brand of the product
